Some small improvements to File and MathExtra namespaces

+++ include/File.h
@@ -1,3 +1,8 @@
+// Namespace for file processing.
+// Joris De Ridder & Enrico Corsaro @ IvS - 12 February 2013
+// e-mail: joris.deridder@ster.kuleuven.be
+// Header file "File.h"
+// Implementation contained in "File.cpp"
 
 #ifndef FILE_H
 #define FILE_H
@@ -16,11 +21,10 @@ using Eigen::ArrayXd;
 
 namespace File
 {
-    ArrayXXd arrayFromFile(ifstream &inputFile, int Nrows, int Ncols, char separator = ' ', char commentChar = '#');
+    ArrayXXd arrayFromFile(ifstream &inputFile, unsigned long Nrows, int Ncols, char separator = ' ', char commentChar = '#');
     void arrayToFile(ofstream &outputFile, ArrayXXd array, string separator = "  ", string terminator = "\n");
     void arrayToFile(ofstream &outputFile, ArrayXd array1, ArrayXd array2, string separator = "  ", string terminator = "\n");
-    void snifFile(ifstream &inputFile, int &Nrows, int &Ncols, char separator = ' ', char commentChar = '#');
-}
-
+    void snifFile(ifstream &inputFile, unsigned long &Nrows, int &Ncols, char separator = ' ', char commentChar = '#');
+} // END namespace File
 
 #endif
